+++ packages/kernel/src/validation.ts
