+++ src/commands/recording/runtime/recording.ts
@@ -8,7 +8,7 @@ import type { ArtifactDescriptor, FileOutputRef } from '../../../io.ts';
 import type { CommandContext } from '../../../runtime-contract.ts';
 import { AppError } from '@agent-device/kernel/errors';
 import { successText } from '@agent-device/kernel/success-text';
-import { requireIntInRange } from '../../../core/validation.ts';
+import { requireIntInRange } from '@agent-device/kernel/validation';
 import {
   recordingQualityInputToExportQuality,
   type RecordingExportQuality,
