+++ src/daemon/app-events.test.ts
@@ -1,6 +1,6 @@
 import { test } from 'vitest';
 import assert from 'node:assert/strict';
-import { parseTriggerAppEventArgs } from '../app-events.ts';
+import { parseTriggerAppEventArgs } from './app-events.ts';
 import { AppError } from '@agent-device/kernel/errors';
 
 test('parseTriggerAppEventArgs validates event name format', () => {
