+++ src/daemon/app-events.ts
