+++ src/daemon/payload-input.ts
