+++ src/daemon/touch-runtime.ts
@@ -19,9 +19,9 @@ import { resolveClickButton } from '@agent-device/contracts/click-button';
 import type { DeviceInfo } from '@agent-device/kernel/device';
 import { AppError } from '@agent-device/kernel/errors';
 import type { Point } from '@agent-device/kernel/snapshot';
-import { readFillBackendResult } from '../core/fill-backend-result.ts';
+import { readFillBackendResult } from './fill-backend-result.ts';
 import { successText } from '@agent-device/kernel/success-text';
-import { requireIntInRange } from '../core/validation.ts';
+import { requireIntInRange } from '@agent-device/kernel/validation';
 import type { DaemonCommandContext } from './context.ts';
 import type { DirectIosSelectorTarget } from './direct-ios-selector.ts';
 import type { DaemonFailureResponse } from './response.ts';
